Added additional script for upload on forks
This script is for forks to upload their apks to their apk branch for danger to fetch them.
The script which uploads builded apk from travis to the apkbranch in the main repository is
upload-apk.sh. This script was build so that in every PR the apk can be attached as comment
via danger. The use of using two seperate scripts is to ensure that the token doesnot needed
to be changed after every pull request merge. Travis will call this script if not in the main
repository i.e, phimpme-android.