Adds danger to the repository (#170)
Adds danger with android apk links. Now danger will comment it automatically.

Closes #61.

This also adds a script which is for forks to upload their apks to their apk branch for danger to fetch them. The script which uploads builded apk from travis to the apkbranch in the main repository is
upload-apk.sh. This script was build so that in every PR the apk can be attached as comment
via danger. The use of using two seperate scripts is to ensure that the token doesnot needed
to be changed after every pull request merge. Travis will call this script if not in the main
repository i.e, phimpme-android.

This commit would allow easy moderation and testing improving the apps development process as the counter is increasing at an alarming rate.