fixes the crashing of app due to orientation change. (#2382)
This makes the Activity(EditImageActivity) to retain it's state even after rotation and hence prevents the app from crashing while inserting text to the image.
In the AddTextFragment.java file the mTextSticketView was getting null
due to restarting of the activity due to which the app crashed. This fix
allows the activiy to retain its state after rotation and hence the
mTextStickeerView is no longer null.