Fixed Image Delete Issue , TrashBin Issue and More
Clicking on Delete Icon on the Toolbar now deletes all the images.
Fixed issue when images of same name are added by changing the PrimaryKey from binPath to an auto_increment id.
Fixed an issue in which files are not deleted from ".nomedia" directory.
Also Updated some external library versions.

Fixed Image Delete Issue , TrashBin Issue and More

Clicking on Delete Icon on the Toolbar now deletes all the images.
Fixed issue when images of same name are added by changing the PrimaryKey from binPath to an auto_increment id.
Fixed an issue in which files are not deleted from ".nomedia" directory.
Also Updated some external library versions.

Update build.gradle

Origin

NullFix and Search Fix

Images stored in Root directory or /sdcard now won't show or store null value as name.
As a result on search and sorting ,
there would be no crash or termination since we are giving a name to the folder incase of that image is stored in /sdcard.

Done with gradlew SpotlessApply