Fixed Image Delete Issue , TrashBin Issue and More
Clicking on Delete Icon on the Toolbar now deletes all the images.
Fixed issue when images of same name are added by changing the PrimaryKey from binPath to an auto_increment id.
Fixed an issue in which files are not deleted from ".nomedia" directory.
Also Updated some external library versions.